Admin UI changes
In ManageEquipment the admin is prompted with a choice for either trucks or trailers, when selected the appropriate table will populate.
In ManageUsers, removed the password column since it is never retrieved for security.
In both Manage pages the delete function has been changed to deactivate, or activate if the object is already deactivated.