NotificationIcon
I added another field in the notification model to accept the incident id so the front end can render the proper pop up when clicking the notification. Also added a max length to the notification list before a scroll bar is introduced. While updating the models i changed the logid to autoincrement. Finally, restored some of the filter options in the logs page.